Update Client.java
I have made a number of changes. These include changing .read to .readLine at line 110, adding \n to all the messages sent to the server and I have changed the Auth to reflect the username of my system. All of these changes were necessary for the client to pass the provided tests. I also pre-initialised the jobData string array to ensure that I don't get out of bounds exceptions.